requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via starlette
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
aws-xray-sdk==2.12.1
    # via okdata-agresso (setup.py)
botocore==1.34.11
    # via aws-xray-sdk
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
deprecation==2.1.0
    # via python-keycloak
ecdsa==0.18.0
    # via python-jose
idna==3.6
    # via
    #   anyio
    #   requests
jmespath==1.0.1
    # via botocore
jsonschema==4.20.0
    # via okdata-sdk
jsonschema-specifications==2023.12.1
    # via jsonschema
okdata-aws==2.0.0
    # via okdata-agresso (setup.py)
okdata-sdk==3.1.0
    # via
    #   okdata-agresso (setup.py)
    #   okdata-aws
packaging==23.2
    # via deprecation
pyasn1==0.5.1
    # via
    #   python-jose
    #   rsa
pydantic==2.5.3
    # via okdata-aws
pydantic-core==2.14.6
    # via pydantic
pyjwt==2.8.0
    # via okdata-sdk
python-dateutil==2.8.2
    # via botocore
python-jose==3.3.0
    # via
    #   okdata-sdk
    #   python-keycloak
python-keycloak==3.7.0
    # via okdata-sdk
referencing==0.32.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   okdata-agresso (setup.py)
    #   okdata-aws
    #   okdata-sdk
    #   python-keycloak
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via python-keycloak
rpds-py==0.16.2
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.0
    # via anyio
starlette==0.34.0
    # via okdata-aws
structlog==23.3.0
    # via okdata-aws
typing-extensions==4.9.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==1.26.18
    # via
    #   botocore
    #   okdata-sdk
    #   requests
wrapt==1.16.0
    # via aws-xray-sdk
